+++ CHANGES.rst
@@ -17,6 +17,15 @@ Breaking:
   [pnicolli]
 - Handled aria-expanded and ESC keypress event via ES [nzambello]
 
+1.2.0 (unreleased)
+------------------
+
+- Remove unused old code
+  [cekk]
+- Fix menu render when there are no settings for the current folder
+  [cekk]
+- Add test suite (functional and acceptance)
+  [cekk]
 
 1.1.0 (2018-05-15)
 ------------------